Don’t crash if sprockets is absent
Currently Webpacker has an implicit dependency on Sprockets. It enhances `assets:precompile` so trying to run any rake task will fail if webpacker is in the Gemfile but Sprockets hasn’t been initialized.

A good next step might be to define `assets:precompile` to `webpacker:compile` if the former is absent, for compatibility with existing tooling.