Allow Rails 4.x support
Technically, nothing in webpacker is dependent on Rails 5.x.
Although there is only official support for Rails 5.1, this
minor change allows the usage of webpacker for Rails 4.x. This
is helpful for enterprise users of Rails that are still working
on supporting Rails 5.x but really want to use the new conventions
for supporting Webpack before fully upgrading to Rails 5.1.