Need to called load_instance for manifest in find!
Since load_instance is a no-op for production, this should be fine
performance-wise.

If we're using `webpack -w` to recompile when files changes, we need
to call this before find!

Otherwise, we don't get new files.

So if we startup both a compile process and the server with a Procfile,
then Webpacker will load with no values in the manifest, and it will not
be read again.

* In the case of using