Do not force SASS indented syntax in sass-loader
According to https://github.com/sass/node-sass#indentedsyntax, `indentedSyntax` forces `sass-loader` to use SASS indented syntax.
The loader is currently configured to process both SASS and SCSS, but this setting prevents SCSS files to be used, as it results in a syntax error (because the loader tries to process it as SASS).