Hash files based on output rather than source
By default webpack hashes files based on their source contents, when doing any
sort of uglification or transpiling this allows the output to change without the
digest changing.

This is a known issue(https://github.com/webpack/webpack/issues/4659) and has an
actively maintained plugin written specifically to alleviate that issue
(https://www.npmjs.com/package/webpack-plugin-hash-output)

This patch set adds a dependency on `webpack-plugin-hash-output` and adds it to
the default plugin list so that any projects which depend on webpacker's
environment will have properly generated hashes that will work with caching.