Fixes Compiler.env accessible in binstubs (#1658)
but these are only set when using the compiler directly (e.g. via
`assets:precompile`). Environment variables set via the compiler are not
accessible to `./bin/webpack` or `./bin/webpack-dev-server`, so manual
compilation and development access does not work.

This fixes #1641. Webpacker::Compiler.env will be accessible in the
binstubs.