allow dev runner to behave like main runner
We allowed the use of ARGS in `Webpacker::WebpackRunner` to pass
down any necessary arguments to webpack itself:

```ruby
cmd = [ "#{@node_modules_path}/.bin/webpack", "--config", @webpack_config ] + @argv
```

However, this was not the case for `Webpacker::DevServerRunner`. Hence,
it was not possible to run the dev runner as such:

```bash
./bin/webpack-dev-server --inline=false
```

This is depite the fact that ARGS are passed to DevServerRunner.