Run binstubs based on Rails.root
This changes the assumption that these tasks are being run in the rails
root to explicitly determine the binstub directory based on the Rails
root.

This fixes the issue for when the tasks are being run from a mounted
dummy app as part of a Rails engine.

Given a fresh Rails plugin: `rails plugin new rails-plugin --mountable`
that has webpacker gem installed.

Before this change running running `app:webpacker:check_binstubs` would
check in `./bin` for webpack binstubs whereas they are actually expected
in `test/dummy/bin`.

After this change the binstubs for the rails app in `test/dummy/bin` are
checked as expected.