Allow dependant tasks to run under a prefix
If you have a scenario where a Rails apps tasks are placed under a
namespace these tasks that invoke other tasks don't work.

The common scenario where this happens is when a Rails engine is mounted
in a dummy app as per: https://github.com/rails/rails/blob/master/railties/lib/rails/tasks/engine.rake#L4-L17

This change resolves the issues where you receive an error of task not
found when running an app command e.g. app:assets:precompile.

If we are able to know a task dependency prior to runtime we don't have
to do anything complex as per the change for `clobber` since it being
defined outside a block does not require a namespace.

However for the tasks that we need to perform actions at runtime we have
to work out the prefix of the task that was executed to then re-use this
on the tasks we want to execute.