Run yarn tasks within the Rails root directory
This resolves problems that occur when the expectation is that the
Rails.root is the same as the working directory of the rakefile.

A common situation for this is when you have a mounted Rails engine in
the test/dummy app to run your engine.

This changes the yarn commands to run in the directory where you would
have the package.json for the rails app that would be used for mounting
rather than the root of the engine. And the node_modules are also
installed within that directory so they are in the expected place for
tasks such as app:assets:precompile.