Update the gitignore within the rails root
The contents of the gitignore only really make sense for the root
directory of a Rails application. If we are dealing with a Rails engine
that mounts the engine onto a test one for usage then it does not make
sense to update the root .gitignore and if someone does have one in the
`test/dummy` directory then that one should be updated instead.