Run webpacker:info in Rails.root
This is resolution to problems that occur with a mounted Rails engine.
With one of these you would want to run `rake app:webpacker:info` and
have it run against the Rails app you'd be mounting in (normally in
test/dummy) however it would incorrectly run in the root directory -
where you probably don't want to have webpacker and webpacker-dev-server
installed.