Basic tests for a mounted application
This sets up a barebones mounted Rails engine so that tests can be
performed against it.

This follows a similar approach to the other rake tasks test and thus
doesn't intend to be too exhaustive in what is testing and is mostly a
cursory test that one of the tasks does perform differently in a mounted
Rails engine context.