Run yarn check --verify-tree as part of the yarn integrity check
From the yarn documentation[1]:

`yarn check --integrity` checks that versions and hashed
values of the package contents in the project’s package.json
match those in yarn’s lock file.

That doesn't actually check that the packages are actually
installed in node_modules. To check that we need to run
`yarn check --verify-tree`

[1]: https://yarnpkg.com/lang/en/docs/cli/check/