Revert "Remove context from file loader (#1914)"
This reverts changes to how assets are managed introduced in RC6. The
changes themselves were introduced to fix #1887 but I couldn't draw a
line between the issue and how these changes were meant to address it.
Additionally, they didn't appear to help the user who made the initial
request and caused a host of other problems.

This appears to be the potential cause for #1938, #1845, #1915, #1947,
and #1922.

This reverts commit 35ec3aee4fd1cf4bb9fa5303fe83a5838b9d9409.