Temporarily fix Travis CI builds
`nvm install node` will install the latest node version, which is 12 as
of writing. Because of a problem in the resolved node-sass version this
makes the builds fail.

To get green builds until a proper fix for node 12 is introduced we'll
install node 10, which is a LTS release, while 11 will be EOL in June
2019.

@rokumatsumoto is maybe working on a fix, but I think green builds,
especially for third party PRs are an important intermediate step.

https://github.com/rails/webpacker/issues/2077
https://github.com/sass/node-sass/pull/2633
https://github.com/nodejs/nan/issues/849
https://github.com/nodejs/Release