Update node-sass troubleshooting information
Probably obvious to people who are familiar with the tools, I wasn't
sure how to fix the `node-sass` error as described. Hopefully these
changes save some other folks some of the time it took me to dig into
the sassy issue.

Updates the docs to with two bits of information:

1. Guidelines to move any webpack-related packages from
   `devDependencies` to `dependencies`
2. The instructions to rebuild `node-sass` if needed

Was able to verify this works with Rails 6.0.0