De-symlink paths in babel loader to match Webpack
Webpack passes absolute, de-symlinked file paths to loaders. Presently the babel loader uses resolve(), which does not de-symlink file paths.

In the case of a path (likely from `resolved_paths`) being a symlink including /node_modules/ but the de-symlinked path not including /node_modules/, both the babel loader and the nodeModules loader will skip the file, potentially causing errors because babel transpilation is skipped.

One use case for setting `resolved_paths` this way is using something like `yarn link` to edit a local copy of a package.

This commit modifies the babel loader to de-symlink paths from `resolved_paths` using `realpathSync()` so that those paths match the already de-symlinked paths passed into the loader by webpack. In the case of a non-existent path or other error, it falls back to the prior behavior of `resolve()`.