Check if request path starts with '/packs/' instead of '/packs'
Only checking if it starts with '/packs' can lead to false positives where it forwards requests to
webpack-dev-server that don't strictly match the config.public_output_path and which can't be
handled by the currently running webpack-dev-server — even requests that don't even have anything to
do with webpacker but just happen to start with '/packs'.

Better to pass on the request to Rails unless it is a definite match for the public_output_path that
webpacker-dev-server is responsible for.