Document which yaml config used by Ruby vs. JS
Some of the yaml settings matter only for the Ruby parts of
rails/webpacker. Others are needed for the JS parts of building the
bundles via rails/webpacker.

If one's app is using a custom webpack build, then the non-ruby settings
can be ignored. On the other hand, the ruby settings are critical to all
Rails apps.

Furthermore, when using a custom RAILS_ENV such as something like
"production_local," then it's critical that all the Ruby values be set
as defaults.