Allow watched_paths to be configured in YAML file
Allows for a `watched_paths` key to be configured in the YAML
configuration file.

There is a change in behaviour between using the YAML key and
`Webpack::Compiler.watched_paths` as the YAML key will go through the
globbing step in the same way that `resolved_paths` does, whereas using
`Webpack::Compiler.watched_paths` does not.

Not sure whether that subtle change is ok, can remove the globbing if
that is preffered.