Allow configuration of the WEBPACKER_CONFIG via env
If you're using the rails/webpacker node package from an installation
where the node_modules is a subdirectory of the main rails directory,
then the package has an error reading the main config/webpacker.yml.

This is because config.js and env.js assume (maybe incorrectly) that the
node_modules directory is at the top level of the rails project.

Instead, an installation may like the node_modules for the frontend or
client to be in a directory named as such.

This change sets the ENV value in the 3 places that webpack is called:

1. bin/webpack
2. bin/webpack-dev-server
3. compiler called from within the Rails code when the bundles are
   stale.