Always use content hash for MiniCssExtractPlugin
With Javascript packs in development mode, files are outputted with
hashes in the file names so browser cache is automatically busted when
code is changed and new files are made.

However, the mini-css-extract-plugin plugin is only configured to output
files with hashes in production mode. This means that any changes to CSS
files locally will appear to not have any effect on the page you are
working on because the browser is assuming the resulting CSS file with
the same name has not changed and reloads the cached version, forcing us
to manually bust browser cache to see our changes.

This code change enables the hash-based file names for the plugin
regardless of environment.