Clarify dev_server.running? first checks configured
Webpacker.dev_server.running? might appear to be costly, but it first
checks that the dev_server value is on the config Hash.

Renaming the method makes it clear that the first check is whether it is
configured.

Since the dev_server would ever be configured on production, there is no
cost for calling this method on production.