Upgrade Jest
ran the following to upgrade/update jest: `yarn upgrade jest`

This seemed to resolve a lot of the `yarn audit` high warnings pertaining to
lodash.

If you perform a `yarn audit --level high` on 5-x-stable and the same command
on this branch you should see a lot of the yarn audit security warning go away.

Please let me know if there is any reason to avoid this upgrade or potentially
and edge case that maintainers of v5 webpacker have ran into in the past.