Update webpack-dev-server config for version 4
`webpack-dev-server` 4 dropped very recently. With it came a very large set of breaking changes regarding the configuration.

This updates the dev server configurtation in `development.js` and the default `webpacker.yml` files to work with it.

I can at least confirm that using this against a relatively basic Rails app using version 4 with tailwind + turbo works as "expected."