webpacker:install does not require webpacker.yml
If you run ./bin/rails webpacker:install on a rails app with no
config/webpacker.yml, we get a crash with this error:

> Webpacker configuration file not found webpacker-v6/config/webpacker.yml.

> Please run rails webpacker:install Error: No such file or directory @
rb_check_realpath_internal - webpacker-v6/config/webpacker.yml

With this fix, allow installer to run.

Problem solved with a global to indicate installing so that default
configuration is used rather than crashing because the yet to be
installed config/webpacker.yml is not yet installed.